requirements.txt
<<<<<<< HEAD
# Iris AI Stream Companion - Dependencies
# Полностью бесплатные технологии

# TTS (синтез речи)
edge-tts>=7.0.0
pygame>=2.5.0

# Speech Recognition (распознавание речи)
vosk>=0.3.45
SpeechRecognition>=3.10.0

# AI (генерация ответов)
groq>=0.5.0

# Web/API
flask>=3.0.0
websocket-client>=1.6.0
websockets>=12.0
requests>=2.31.0

# Utils
python-dotenv>=1.0.0
numpy>=1.24.0

# Audio (опционально, для Windows)
# pycaw>=20230407  # Windows only
# sounddevice>=0.4.6  # требует PortAudio
=======
groq==0.9.0
pygame==2.6.1
pyttsx3==2.90
openai==1.12.0
numpy==1.26.4
python-dotenv==1.0.0
>>>>>>> 6d0ea0cd1396a0d7b9b7fabfc564c9750f26d5aa
